requirements.txt
# =============================================================================
# ZarrinGold Django Backend — Python Dependencies
# =============================================================================
# Core Framework
django>=4.2

# REST API
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3

# CORS
django-cors-headers>=4.0

# Database
psycopg2-binary>=2.9

# Filtering
django-filter>=23.0

# WebSocket / Real-time
channels>=4.0
daphne>=4.0
channels-redis>=4.0

# Environment variables
python-decouple>=3.8

# Image processing
Pillow>=10.0

# Storage backends (S3, etc.)
django-storages>=1.14

# WSGI HTTP Server
gunicorn>=21.0

# JWT Blacklist (for token rotation)
djangorestframework-simplejwt[crypto]>=5.3
